backend/requirements.txt
fastapi~=0.115.7
uvicorn~=0.34.0
aerich~=0.7.2
asyncpg~=0.30.0
tortoise-orm~=0.21.7
bcrypt~=4.2.0
passlib~=1.7.4
python-jose~=3.3.0
python-multipart~=0.0.17